% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propyrs.R
\name{capatplot}
\alias{capatplot}
\title{Plots cause of death pattern for a region over time}
\usage{
capatplot(ag, ctry, cas, hltf = NA, aws = NA, alabs = agelabs, ca2 = "all")
}
\arguments{
\item{ctry}{WHO region code (use "all" for all regions).}

\item{cas}{Vector of cause codes.}

\item{hltf}{Data frame in the format returned by hmdltfs.
If defined, plot expected proportion of deaths (the caldx column) for someone at age ag.}

\item{aws}{Data frame with age groups for aggregation.
The standard frame "aw" is suited for causes related to aging, with higher age groups fine-grained.
The frame "aw_y" is suited for causes commong among young adults, like HIV and external causes.}

\item{alabs}{Vector with labels of the same length as the age groups.}

\item{ca2}{Denominator cause (use "pop" for population to get death rates).}

\item{age}{Age group.}
}
\description{
Plots cause of death pattern for a region over time
}
\examples{
capat <- c("ihd", "othhd", "othath", "othcirc", "str", "neurdegnovd", "diab",
"chresp", "inf", "othdis", "illdef", "ext", "tum")
capatplot(1, 4290, capat)
}
