% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propyrs.R
\name{ctry_awsyplot}
\alias{ctry_awsyplot}
\title{Plots cause of death ratios for different age groups in a region}
\usage{
ctry_awsyplot(
  ctry,
  ca1,
  ca2,
  minyr,
  maxyr,
  fwscales = "fixed",
  ycol = "rat",
  aws = aw,
  alabs = agelabs_w
)
}
\arguments{
\item{ctry}{WHO region code (use "all" for all regions).}

\item{ca1}{Numerator cause.}

\item{ca2}{Denominator cause (use "pop" for population, to get death rates).}

\item{minyr}{Minimum year to be plotted.}

\item{maxyr}{Maximum year to be plotted.}

\item{fwscales}{Using free or fixed scales for the plot.}

\item{ycol}{Column to use for the plot. Use "meanrat" for mean of death rates over age groups.}

\item{aws}{Data frame with age groups for aggregation.
The standard frame "aw" is suited for causes related to aging, with higher age groups fine-grained.
The frame "aw_y" is suited for causes commong among young adults, like HIV and external causes.}

\item{alabs}{Vector with labels of the same length as the age groups.}
}
\description{
Plots cause of death ratios for different age groups in a region
}
\examples{
ctry_awsyplot("4290", "ihd", "all", 1951, 2023)
}
