% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propyrs.R
\name{ctry_caf}
\alias{ctry_caf}
\title{Generate a data frame with cause of death ratios over time for one or more regions}
\usage{
ctry_caf(ctry, ca1, ca2)
}
\arguments{
\item{ctry}{WHO region code (use "all" for all regions).}

\item{ca1}{Numerator cause.}

\item{ca2}{Denominator cause (use "pop" for population, to get death rates).}
}
\description{
If a CSV file with matching name exists in the extdata subdirectory, it will be used.
Otherwise, the result will be saved to such a file, which can be re-used to avoid
the relatively computing-intensive calculations with regular expressions on the
WHO data files.
}
\examples{
ctry_caf(4290, "ihd", "all")
}
