% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propyrs.R
\name{thrctries}
\alias{thrctries}
\title{Returns a data frame of countries with minimum and maximum years with a certain cause ratio over a threshold.}
\usage{
thrctries(caf, thr, nthr, tsex, tage)
}
\arguments{
\item{caf}{Frame returned byt ctry_caf.}

\item{thr}{Include countries with ca1/ca2 ≥ thr.}

\item{nthr}{Include countres with ca2 ≥ nthr (to avoid small populations which can give very noisy results).}

\item{tsex}{Sex to include.}

\item{tage}{Age group to include.}
}
\description{
Returns a data frame of countries with minimum and maximum years with a certain cause ratio over a threshold.
}
\examples{
circall <- ctry_caf("all", "circ", "all")
circall |> thrctries(0.5, 200, 2, 1)
}
